Remove PartialOrd and Ord implementations
The vectors for which a partial ordering exists
have a `vec::partial_ord(&self)` method that
returns a wrapper that implements at least `PartialOrd`.
These wrapper implements `Ord` when possible.

The vectors for which a total ordering exists have
a `vec::ord(&self)` method that returns a wrapper
that implements `Ord`.

Closes #48.