Fix undefined behavior reported by miri when reading or writing slices (rust-lang/packed_simd#351)
Many methods in the slice or array api take a pointer to the first
element and then use that pointer to read or write multiple lanes.

This is undefined behavior according to miri since pointer is only valid
for a single element. The fix is to use the `as_ptr`/`as_mut_ptr`
methods of the slice, which gives a pointer to the whole slice and also
leads to nicer code.